requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docopt==0.6.2
enum34==1.1.10
idna==2.10
importlib-metadata==1.6.1
jmespath==0.10.0
lxml==4.5.1
more-itertools==8.3.0
namedlist==1.7
packaging==20.4
paho-mqtt==1.3.1
pbr==5.4.5
pluggy==0.13.1
py==1.8.1
PyJWT==1.7.1
pykwalify==1.7.0
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-rerunfailures==9.0
python-box==4.2.3
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
stevedore==2.0.1
tavern==1.2.2
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
